requirements.txt
numpy 
discord.py == 1.7.2
xkcd == 2.4.1